Move secondary editors to their own files
